requirements.txt
flask>=3.0
flask-socketio>=5.3
pyserial>=3.5
evdev>=1.7
python-dotenv>=1.0
